+++ app/src/main/java/com/example/vitabu/WriteReviewActivity.java
@@ -161,17 +161,17 @@ public class WriteReviewActivity extends AppCompatActivity {
         // update the user's rating
         String username = user.getUserName();
         if (username.equals(review.getOwnerName())){
-            int rating = user.getOwnerRating();
+            float rating = user.getOwnerRating();
             int numReviews = user.getNumOwnerReviews();
             rating = (rating*numReviews + review.getRating())/(numReviews+1);
             user.setOwnerRating(rating);
             user.setNumOwnerReviews(numReviews +1);
         }
         else{
-            int rating = user.getBorrowerRating();
+            float rating = user.getBorrowerRating();
             int numReviews = user.getNumBorrowerReviews();
             rating = (rating*numReviews + review.getRating())/(numReviews+1);
-            user.setBorrowerRating((int) rating);
+            user.setBorrowerRating(rating);
             user.setNumBorrowerReviews(numReviews +1);
         }
